requirements.txt
anyio==3.6.2
attrs==22.2.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pluggy==1.0.0
Pygments==2.14.0
pytest==7.2.2
PyYAML==6.0
rich==12.6.0
shellingham==1.5.0.post1
sniffio==1.3.0
tomli==2.0.1
typer==0.7.0
watchfiles==0.18.1
